django_backend/requirements.txt
Django==3.2.3
django-crispy-forms==1.11.2
djangorestframework==3.12.4
Jinja2==2.10.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
virtualenv==20.4.7